vm: Implement page promotion and demotion
Now that the codebase is ready for TVMs to be mapped with huge pages,
both page promotion and page demotion functionality are implemented.

This extends the TEE-Host API, with TvmPromotePage allowing the host
to turn 512 pages of a certain size into one huge page of the next
larger size. And with TvmDemotePage allowing the host to split one page
into 512 pages of the next smaller size.

For promotion, that means the PTE at the previous page level is promoted
as a leaf and the page table page used for the 512 subpages is simply
cleaned and freed. The new leaf is set with the PFN of the huge page.

For demotion, the PTE at the page level is turned into a non-leaf and it
points to the new page table page that is needed to refer the new 512
page entries. Every entry (unused so far) must be turned into a leaf.

Signed-off-by: Sebastien Boeuf <seb@rivosinc.com>