vm: Update existing functions to support all page sizes
In particular, updates all guest_add_*_pages() functions so they can
support pages of all sizes. This relies onto various functions from
vm_pages.rs which have been modified as well so they don't assume 4k
pages.

It's important to note that both do_get_converted_pages() and
get_shareable_pages() process the pages iterator with a filter that
ignore pages that are not aligned on the expected page size. This is
because converted pages and shareable pages come from the host, where
they are mapped as 4k pages. The filter skips intermediate pages so that
a LockedPageList with pages of the right size is created. This is only
possible because we previously validate the range is contiguous and
suitably aligned.

Signed-off-by: Sebastien Boeuf <seb@rivosinc.com>