host_vm: Align GPA/SPA in 2M chunks
In order to support 2M pages being used in the TVMs, the host must have
been previously mapped with 4k pages that are 2M-aligned chunks.

The root page table requirement to be 16k-aligned still stands but for
zero pages and measured pages, we must ensure it's now 2M-aligned
chunks. The HwMemMap is now built with this requirement so that kernel,
initramfs and fdt can be allocated on 2M-aligned regions.

Signed-off-by: Sebastien Boeuf <seb@rivosinc.com>