Salus: Switch stack to VA mapped.
This patch allows both bootstrap CPU and secondary CPUs to switch to
VA mapped stack once the pagetable is set (before calling `main`).

PerCpu data is now allocated in a separate array and it is separate
from stack for all CPUs. This avoids allocating a full extra stack for
the bootstrap CPU just to use as PerCpu.

The setting of TP register to the PerCpu data pointer is now explicit
also for the secondary CPUs, as it cannot be passed at start.

This patch also requires all stack for all CPUs to be of the same size.

This patch incidentally fixes quite a few corruption bugs. We were
running out of stack with 8 pages previously, but it was difficult to
catch on a physically mapped stack.

Signed-off-by: Gianluca Guida <gianluca@rivosinc.com>