ci: Build each commit in the forward order
By default, git rev-list returns a list of commits in reverse order.
This caused an unexpected behavior with the tellus/guestvm integration
tests. Because the last commit to be built was the first commit of the
pull request, both salus and tellus_guestvm binaries didn't include all
the changes from that same pull request.

The way to fix this is by using --reverse option from git rev-list, so
that we can walk through each commit in the forward direction.

Signed-off-by: Sebastien Boeuf <seb@rivosinc.com>