riscv-page-tables: Do not walk for each PTE on lock.
Previously, we were walking the page table for each pte. This was
slow.

This commits find the page-table at a certain level, and then iterates
over the indices. In case we reach the end of the page-table but we're
not done yet, we walk again, and start from the beginning of the
page-table.

This speeds up the locking operation of PTEs quite a lot.

Signed-off-by: Gianluca Guida <gianluca@rivosinc.com>