page-tracking: Get rid of Unassigning state
Relying on the new block and remove flow, we can get rid of the
Unassigning state which was nothing but a duplicate of what is handled
through this new design.

The IMSIC file unbind/rebind codepath was the one still using the
unassignment method, which is now replaced with the block and remove
equivalent.

Signed-off-by: Sebastien Boeuf <seb@rivosinc.com>