imsic: Support sparse imsic regions
If m-mode interrupt files are interleaved with S and guest files, then
the device tree may report holes in the reg map for the imsics, with
each cpu's files separated by a missing page.

Secondary reg regions will have non-zero hart ids, so remove the check
that all regions are for hart 0.

Similarly the per_hart_size is more of a stride and it's possible to
have less than that size in each region.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>