ci: github - submodule update before build
there is a bug where a PR updates a submodule in a commit other than the
first in the chain. The tree will have the updated submodule from the
later commit instead of the one that was current for the commit being
tested. This is because the CI script didn't update the submodules
before testing each commit.

Add the update so that each commit is tested with the submodules it
specifies.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>