drivers/iommu: Device directory to support both base and extended format
Generalize the device directory implementation to allow using either
base or extended device context format. At the table implementation
level, there is now a type parameter for the device context that we're
using, along with layout information and member accessors in a trait
implemented for both the `DeviceContextBase` and `DeviceContextExtended`
variants. The API remains agnostic of the type parameter, and we're
instead selecting the underlying table implementation type to use via a
run time parameter, with a new `DeviceDirectoryOps` trait and the
`enum_dispatch` crate helping to bridge from the type-agnostic layer to
the typed table implementation.