Implement SBI DebugConsoleFunction::Read support
A basic implementation analogous to the write implementation that reads
characters from the UART. This has been tested successfully against with
the Linux SBI hvc driver. While at it wrap access to the buffer in a
`PinnedBuffer` helper that isolates the business logic from safety
considerations by exposing the buffer as a `VolatileSlice`.