Add `hardware_ad_updates` build flag
The new build flag indicates whether hardware support for updating A/D
bits in PTEs should be used. This is enabled by default and can be
adjusted by passing `--//:enable_hardware_ad_updates=false` to the bazel
invocation. The flag gets reflected into a cargo feature, which
respective crates inspect. In hardware A/D update mode, the `svadu` CPU
extension and the `AMO_HWAD` IOMMU capability must be present. When
built with `hardware_ad_updates` disabled, PTEs are initialized with the
A/D bits set, side-stepping the need for hardware updates and thus not
requiring `svadu` and `AMO_HWAD`.