drivers/iommu: Probe device directory mode
IOMMU implementations are not required to support all device directory
modes. The way for software to determine whether a mode is supported is
to attempt to program the mode and read back the DDTP register to see
whether the mode value was accepted.

This change replaces the hard-coded 3-level mode with a probe loop to
attempt mode values. The loop tries modes in decreasing number of levels
to maximize the number of devices that can be managed in the table.